Pull fuzzing FakePayloadDigest out into own module for now
