+++ fuzz/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod encode;
 pub mod path;
+pub mod placeholder_params;
